original and default model the same
The problem in the previous commit was that Danielas methods
implicitly use the effects from the previous sort that shuffled
instances in place.

However ...
In the classification case, Daniela's original version has some bug
that causes wrong number of instances in the leaves of a tree, hence
the output files do not match exactly.